Logic linreg

+++ include/adc_driver.h
@@ -10,6 +10,6 @@
 #include "config.h"
 
 void adc_setup();
-void adc_read(int *value);
+void adc_read(double *value);
 
 #endif /* __ADC_DRIVER_H__ */
