+++ tests/mock/mock_driver.cpp
@@ -7,11 +7,11 @@ void mock_setup(char *file_name)
 	in_file = fopen(file_name, "r");
 }
 
-void mock_read(int *value)
+void mock_read(double *value)
 {
 	char line[6];
 	if (fgets(line, 6, in_file) != NULL) {
-		   sscanf(line, "%d", value);
+		   sscanf(line, "%lf", value);
 	} else {
         *value = -999;
     }
